move position to public, refactor neighborhood tracker constructor
