+++ python/pythonlab/neighborhood/neighborhood/__init__.py
@@ -2,6 +2,7 @@ from .painter import Painter as Painter
 from .neighborhood_log import NeighborhoodLog as NeighborhoodLog
 from .painter_log import PainterLog as PainterLog
 from .painter_event import PainterEvent as PainterEvent
+from .position import Position as Position
 # We export World so we can clear the world in pythonlab_setup and
 # handle setting the context type in unittest_runner
 from .support.world import World as World
