+++ python/pythonlab/neighborhood/tests/support/test_painter_tracker.py
@@ -2,7 +2,7 @@ from neighborhood.support.painter_tracker import PainterTracker
 from neighborhood.support.neighborhood_signal_message import NeighborhoodSignalMessage
 from neighborhood.support.signal_message_type import SignalMessageType
 from neighborhood.support.neighborhood_signal_key import NeighborhoodSignalKey
-from neighborhood.support.position import Position
+from neighborhood.position import Position
 
 def test_initialize_painter_tracker():
     painter_tracker = PainterTracker("painter-1", Position(0,0,'East'), 2)
